backend/requirements.txt
# Archivo de requisitos para el backend de la Red Social Universitaria

# Framework FastAPI
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0

# Seguridad
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Cliente Supabase
supabase>=1.0.0

# Utilidades
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.1
email-validator>=2.0.0
pillow>=10.0.0
python-dateutil>=2.8.2
jinja2>=3.1.2

# Exportación de datos
openpyxl>=3.1.2
XlsxWriter>=3.1.0
reportlab>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
